powerbiWorkspace/requirements.txt
requests
azure-appconfiguration
sys